Added more evolved build script.

+++ build.sh
@@ -1,21 +1,38 @@
 #!/usr/bin/env bash
 
-printf -v DATE '%(%Y-%m-%d)T\n' -1
+TYPE=${1:-"all"}
+
+if [ ${TYPE} = "all" ]; then
+  PULL=""
+else
+  PULL="--pull"
+fi
+
+printf -v DATE '%(%Y-%m-%d)T' -1
 
 echo ${DATE}
 
-MLST_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-mlst
-MLST2_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-mlst2
-CGMLST_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-cgmlst
-NGSTAR_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-ngstar
+if [ ${TYPE} == "all" ] || [ ${TYPE} == "mlst" ]; then
+  MLST_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-mlst
+  docker --pull build --rm -t ${MLST_IMAGE} --build-arg TYPE=mlst .
+  docker push ${MLST_IMAGE}
+fi
 
+if [ ${TYPE} == "all" ] || [ ${TYPE} == "mlst2" ]; then
+  MLST2_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-mlst2
+  docker ${PULL} build --rm -t ${MLST2_IMAGE} --build-arg TYPE=alternative_mlst .
+  docker push ${MLST2_IMAGE}
+fi
 
-docker --pull build --rm -t ${MLST_IMAGE} --build-arg TYPE=mlst .
-docker build --rm -t ${MLST2_IMAGE} --build-arg TYPE=alternative_mlst .
-docker build --rm -t ${CGMLST_IMAGE} --build-arg TYPE=cgmlst .
-docker build --rm -t ${NGSTAR_IMAGE} --build-arg SCHEME=ngstar .
+if [ ${TYPE} == "all" ] || [ ${TYPE} == "cgmlst" ]; then
+  CGMLST_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-cgmlst
+  echo "Building ${CGMLST_IMAGE}"
+  docker ${PULL} build --rm -t ${CGMLST_IMAGE} --build-arg TYPE=cgmlst .
+  docker push ${CGMLST_IMAGE}
+fi
 
-docker push ${MLST_IMAGE}
-docker push ${MLST2_IMAGE}
-docker push ${CGMLST_IMAGE}
-docker push ${NGSTAR_IMAGE}
+if [ ${TYPE} == "all" ] || [ ${TYPE} == "ngstar" ]; then
+  NGSTAR_IMAGE=registry.gitlab.com/cgps/pathogenwatch/analyses/typing-databases:${DATE}-ngstar
+  docker ${PULL} build --rm -t ${NGSTAR_IMAGE} --build-arg SCHEME=ngstar .
+  docker push ${NGSTAR_IMAGE}
+fi
